requirements.txt
amqp==5.1.1
asgiref==3.5.2
autopep8==1.7.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
blis==0.7.8
branca==0.5.0
Brotli==1.0.9
catalogue==2.0.8
celery==5.2.7
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cssselect==1.1.0
cymem==2.0.6
dateparser==1.1.1
dnspython==2.2.1
duckduckgo-search==2.0.2
email-validator==1.2.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
folium==0.12.1.post1
fuzzywuzzy==0.18.0
google-trans-new==1.1.9
GoogleNews==1.6.5
googletrans==3.0.0
greenlet==1.1.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.12.0
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.1.0
kombu==5.2.4
langcodes==3.3.0
langdetect==1.0.9
libretranslatepy==2.1.1
lxml==4.9.1
mailjet==1.4.1
mailjet-rest==1.3.4
MarkupSafe==2.1.1
murmurhash==1.0.8
nest-asyncio==1.5.5
newspaper3k==0.2.8
nltk==3.7
numpy==1.23.2
oauthlib==3.2.1
packaging==21.3
pandas==1.4.4
pathy==0.6.2
Pillow==9.2.0
preshed==3.0.7
prompt-toolkit==3.0.31
pycodestyle==2.9.1
pydantic==1.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.3.2
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
scikit-learn==1.1.2
scipy==1.9.1
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.2.1
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.40
srsly==2.4.4
thinc==8.1.0
threadpoolctl==3.1.0
tinysegmenter==0.3
tldextract==3.3.1
toml==0.10.2
tqdm==4.64.1
translate==3.6.1
tweepy==4.10.1
typer==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
uclassify==1.0.0
urllib3==1.26.12
vine==5.0.0
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
